react: Cut lint rule `react/no-multi-comp`.
In a lot of places we have some tiny helper components that exist
entirely to serve a specific parent component.  Most of those would
be a lot easier to read, understand, and maintain if the helpers
were in the same file where they're used -- less boilerplate code;
less need to cross-reference back and forth between files; and most
of all, the helper component typically implicitly assumes things
about the context it's used in, and that context would be explicit
and ready at hand.

In other cases, we have several closely related small components
(like `TitlePrivate`, `TitleGroup`, `TitleStream`, and friends) that
have a lot of logic in common, and moreover that all sit behind a
common interface as far as the rest of the app is concerned.  When
placed in separate modules, all the internal interfaces between these
components end up being exported, and exposed as global interfaces to
the rest of the app, adding a lot to the total complexity of the app.
If combined in a single file, the family of components would together
expose only a much narrower interface to the rest of the app, in
addition to reducing boilerplate code and duplicate code.

So, in short, our codebase would be greatly improved in a number of
places by combining several React components into the same file.

But we've had an ESLint rule enabled called `react/no-multi-comp`
which complains about doing that.  For the reasons above, this is
completely perverse as a matter of code style.  Turn it off.

(In certain contexts this rule might be justified as a regrettable
workaround for limitations of JS toolchains -- as a way of helping
a build to do effective "tree-shaking", i.e. to leave out code that
isn't being used, which JS toolchains typically aren't able to do
except in units of whole modules.  But we only have one app, which
all our code gets bundled together into; so when we aren't using a
given bit of code, in general we simply delete it.)