emoji: Avoid perverse-`reduce` pattern in new conversion code.
These follow a pattern we have in a number of places in our codebase,
but a bad one: we use `reduce`, but then go and just mutate the
accumulator rather than follow the pure-functional structure `reduce`
is intended for.

For looping through some input and mutating an accumulator, there's a
simpler and more explicit and idiomatic pattern: a loop.  Do that.