alert words: Cut disused init/fetch action creators.
This story is *almost* like the last two commits: we haven't used
either of these action creators since 155df7426, in 2017, and if
we did, it'd be inherently racy.

But there is a twist -- we've gone and abused this action type in a
totally separate action creator which is invoked in a different
circumstance entirely, where the `INIT_...` name isn't appropriate.
(Namely, in `eventToAction` when handling an `alert_words` event
from the server.)  So the action type survives; we'll give it a
better name in the next commit.