tools/test android: Build test files up front, avoiding dupe failures.
When tests fail, we rerun them in order to get the names of failed
tests, as an awkward workaround for limitations in the quiet/verbosity
options of Gradle and the test runner.

But if tests fail to even *build*, we don't need the rerun;
the original failure has perfectly OK output, and the rerun
doesn't have any more information.  So do the build up front,
in the same command where we build the (debug) app itself,
where we don't have that rerun logic.