webview: Add ability to copy links found in messages.
There is no way to copy only links from a message at the moment,
which makes it hard for some users to copy only the links to share
with other people.

There could be two solutions to this.  One is that another menu
option could be provided to copy only the link in your message,
which is not a correct approach as immediately some problems come to
mind.  For example, if there were two links in the message, which
one would the app copy?

Another solution to this could be what commercial applications
already apply, which is if you hold the link text, it gets copied.
Which is what has been implemented here.  I have modified the
already existing `MessageListEventLongPress` to include
functionality for long-pressing on links.

Tests have been conducted on iOS simulator and Android emulator.
Functionality has also been checked for Android on a real device.

This commit also made conspicuous a related bug which already
affected long-presses on iOS (#3429), fixed separately in the
previous commit for the sake of clarity.

Fixes #3385.