message: Make a conditional a bit clearer in `shouldBeMuted`.
It's kind of obscure what the type of `display_recipient` has to do
with our logic here; what we really mean is "is this a PM message or
a stream message".  Say that directly.

The old code also caused ill-typed message fixtures in our tests to
always be interpreted as unmuted, and the new code interprets many of
the same messages as muted.  Fix the tests that break, by making them
something closer to well-typed.

[greg: expanded commit message]