eventMiddleware: Use real types for our assumptions about events.
The `Object` pseudotype disables the type-checker almost completely,
allowing virtually any assumption to get made implicitly about this
value.

The reality is that our assumptions here are not that wild; really
we just assume that `type: 'message'` means a certain set of other
properties and their types, which we've even already written down
under the name `MessageEvent`.  So just use that.