eventToAction: Be more explicit and narrower in disabling type-checking.
This function does in fact encode a large and complex range of
assumptions about the events we get from the server; until we
have a handle on that, we'll be disabling the type-checker one
way or another, and doing it up front in the function's signature
is as good a place as any.

Even that isn't a reason to write the `Object` pseudotype, though
(as indeed there basically never is a good reason to do so).  A
$FlowFixMe type has essentially the same effect on type-checking
(it has exactly the same effect as `any`)... but allows us to
mark the file as `@flow strict-local`, clamping down on non-strict
behavior elsewhere in the file.  Like any use of $FlowFixMe, we
accompany it with a comment about why we've had to settle for it.