nav: Simplify types on `navigation` props.
This form is shorter, and also I think closer to the truth.

On the other hand, as the fixmes show, these types actually cause
the type-checker to give errors when we use these components in
`AppNavigator`!  In fact we'd need fixmes on more of them... if the
react-redux libdef didn't defeat type-checking at `connect` calls.

I think two things are going on in the fixmes:
 * The react-navigation libdef doesn't have the right covariance:
   it ends up wanting us to put more properties into the argument
   to NavigationScreenProp.
 * The libdef doesn't want to guarantee that anything in particular
   goes into `params` -- in which it is entirely correct!  That is
   an unchecked assumption we make, coupling `navActions` with the
   respective screen components.