msglist header: Clarify the logic controlling recipient headers.
The logic for which header to show in which particular narrow is a
little mucky at the moment.  There are two points of data we look at
while deciding this: 1, the type of the message, and 2, the narrow
we are currently in.  The current iteration does not clearly
differentiate between these two, and a developer who is not used to
this part of the codebase might be slightly confused going through
this.  This commit clears up the logic for this file.

The change centers around using `caseNarrow` to find out what kind of
a header is required. For example, if the current narrow type is
`stream`, we would show a header which requires the topic and the date
in the stream.  In the case that the narrow is `home` narrow, we would
show the full recipient header, that is the stream name, topic name,
and the date.  The narrows where the header is not being used as of
this commit are marked as `null`.

We use the marker generated in the previous paragraph to identify the
specific kind of header desired.

Also fix up some tests which pass ill-typed data to this code.