api: Update the `User` type.
It's gained a couple of fields since we last wrote this down.

(The fact that this can change out from under us highlights that it's
not really quite accurate to write the types here, the way we do; a
really robust approach would look more like the "crunchy shell" we've
recently introduced in e.g. FcmMessage.kt, for notification data on
Android.  But that's a bigger project for another time.)