eslintrc: Disable rule react/no-unused-prop-types.
It's true this occasionally points out the next step of a refactor
when you've just removed something.  But really that's an easy check
to make for oneself; and if missed, the consequences probably aren't
yet a bug (not even a latent one).

And on the other hand, sometimes there's a good reason why a
component takes a prop that we don't actually use.

In particular, we define almost all our props types as exact object
types, and many of our components are passed to react-redux's
`connect`... which means they are passed a `dispatch` prop.  That
means if we leave `dispatch` out of the props type, as we often do,
the type is just untrue -- a type error.  With the upcoming real,
effective libdef for `connect`, Flow will become able to spot that
and we'll have to acknowledge `dispatch` in all the applicable
components (or else give up exact object types there.)

More generally, the properties in a React component's props type
play a very similar role to the parameters of a function.  It's
perfectly normal for a function to take some parameters it doesn't
actually use -- this is especially common when the function will be
passed as a callback to some more general interface where other
callbacks might want to use those parameters.  Similarly, we're
passing these components as arguments to the general `connect`
interface, which is designed to also accept components that do need
a `dispatch` parameter.  So it gets passed a value for that
parameter^W prop, whether it wants it or not, and that's fine.