lightbox/shareImage: Fold together split ios/android files.
Fixes #3018 -- this is the last pair!

One of the motivations for #3018 was that the .android.js files
weren't getting type-checked... and this illustrates the problem
nicely.  We had a blatant type error on the `downloadImage` result,
that was just silently ignored; now at least it's explicitly marked
with a fixme as a warning sign.  And we had an `@flow strict`
marking which the file certainly hadn't earned.