connect: Make a trivial wrapper that just narrows down the type.
This just picks one of the N overloads from the libdef -- the only one
we actually need -- and specifies that particular type.  Plus fixes it
up slightly; e.g. the libdef uses the pseudotype `Object`, which
doesn't serve any purpose here that `{}` doesn't do better.

This will be helpful in a bit for isolating changes to the libdef and
to the Flow version from our code, giving us an easy single point of
control for the types Flow sees.