UserItem: Refactor `onPress` to take just `email`.
`UserItem` component and several related ones, when using an
`onPress` callback, were passing an object consisting of `email`
and `fullName` of the user being selected, and a variety of
callbacks used one or the other.

Trying to type this more precisely and 'strict'-ly revealed that
we are treating this object rather loosely.  The `fullName` info
was almost never used.  So, we omit it.

This is a pure refactor that removes the `fullName` information
and keeps the `email` as a string instead of property of an object.

Only `PeopleAutocomplete` is slightly reworked to get the `fullName`
via the `email` value, a trivial `find` in an array of users.