widget: Try to make the "sorry" UI-rendering more helpful.
The term "widget" is internal to Zulip development and won't
necessarily mean anything to the user.  Here's a pass at writing
a message that will give useful information the user understands.

Looking at our user-facing docs, the only discussion of this
feature I see is all about polls:
  https://zulipchat.com/help/create-a-poll
So probably the ideal text here would say this is a poll... except
that isn't very future-proof, and at least on czo it could in fact
be a Tic-Tac-Toe game even today.  So until we have code looking
to see what type of widget it is, let's not be specific about it.

I think "interactive message" should cover pretty accurately what
these are for, and why there's something here you (the user) aren't
seeing.  Also add a line suggesting what to do instead.

And as a small low-effort approximation of the widget, include the
content too: empirically it's like `/poll Where for lunch?` or
`/tictactoe`, either of which is very helpful context for helping the
user at least judge whether what they're missing is something they
care about seeing.