fetch: Revert "Retry fetching messages, indefinitely."
This reverts #3288, aka commit 7ed7e5277 (and so re-opens #3281.)

After this change, Jest would hang after all tests had completed.
(Not sure how I missed that before merge -- did I neglect to run the
tests? >_>)  I noticed the problem a few commits later, and it took a
bit of debugging to identify this test file, and then this commit, as
the culprit.

Some background I found helpful in debugging: facebook/jest#1456 .

The root of the issue seems to be that a lot of these tests fire off
`fetchMessages` actions, or others that in turn invoke that one -- and
then don't wait for them.  Compounding that, many of them are doomed
to fail at their HTTP requests because their fixture states have no
accounts, so there's no realm to base the URLs on.

Anyway, we don't knowingly merge changes that break things.  Merging
this was a mistake; take it back out.  We can merge a version of this
change once we have one that doesn't cause breakage.