msglist js: Make last case in click handler parallel the rest.
Each of these cases is meant to be the end of the matter, so it
should return.  In particular, if we add a new case at the end,
we'll want the existing last case to return so the new case doesn't
fire if it does.  This is all clearer if the last case already ends
with its own `return` statement, just like all the rest do.

If we just add such a statement, ESLint takes it out; which is why
we haven't had one to date.  Disable that ESLint rule for this file.