styles: Round off new HSL color values.
This is being done in order to increase developer productivity
(increase in compatibility in instantly seeing changes to colors
made using a VSCode extension 'Color Highlight'). There are also
cases in these changes where the rounded off color was the desired
color (used in the webapp) in the first place! Eg. BRAND_COLOR isn't
`hsl(169.8, 47.9%, 54.1%)` but is `hsl(170, 48%, 54%)` everywhere in
the webapp.