avatars: Increase size of image to `48px`.
Many chat applications keep their avatar sizes in the message list
big enough to be able to easily distinguish with a glance who the
avatar belongs to. This commit tries to follow that principle. I
have compared the changes to Slack, and they appear to follow the
exact same dimensions.

The rule for padding of `message-brief` elements has also been
edited to align with the content from `message-full`.