webview css: Update values for syntax highlighting from webapp.
This commit essentially takes the colors used in the webapp's SCSS
file for syntax highlighting (`pygments.scss`), and copies that into
the application. This makes the colors between the webapp and the
mobile application exactly the same and hence provides better parity
of experience!

There was an issue while copy-pasting the code from the webapp,
where the background-color on the very corners of a piece of code
written in Night mode was white due to a border-radius (see
https://github.com/zulip/zulip-mobile/pull/3530#issuecomment-507066832),
which is why there is a little diff found in `git diff --no-index
../zulip/static/styles/pygments.scss
src/webview/css/cssCodeColors.js`.
^ the diff in reference:

+    .codehilite {
+        background-color: hsl(212, 28%, 18%);
+    }

Fixes #3524.

[greg: small updates to comment in file]