hidden email: Start getting own-email from /register , not auth.
In previous commits in recent days, we've fixed a number of
spots where we were getting the user's own email directly from
`auth.email` or equivalent, in favor of going through this
common `getOwnEmail` selector.  Now switch the selector to get
it from the proper source, where we've recorded the `email`
returned by `/register`.

Fixes #3196.  Modulo (a) any yet-unspotted bugs, of course, and
(b) handling the transition when the hidden-emails setting (aka
`email_address_visibility`) is turned on or off.  The latter is
expected to be a server-side change.