user: Add jsdoc about the active/deactivated distinction.
This can be pretty confusing -- partly because of some of the names
(perhaps we should rename some of these), and partly because we simply
haven't laid out clear principles for when to use which set of users.
Try to make a start on that, and to warn on the confusing names.