messageActionSheet: Use proper `ownEmail`.
Part of #3196 -- this is an example of code that will break when
hidden-emails are enabled, because the (real) email used for auth
will be different from the (server-generated, fake) email used as
an identifier for the user.

Found by grepping for `auth.email`.  There are still a few more
matches for that pattern, which with one exception (in src/api/)
are also broken; those we'll fix in the next few commits.