eslint: Cut rule flowtype/object-type-delimiter .
This is a formatting rule, which is primarily Prettier's job.  With
the upcoming upgrade (in particular to prettier-eslint, I guess?),
Prettier will start making the opposite choice here.

I don't see a clear reason to override Prettier on this question;
we've just had this rule since a1edcaa16, with no recorded rationale.
So, cut it and defer to Prettier.