prettier: Upgrade to 1.18.2, the latest.
Upgraded like so -- Prettier to the latest, and `prettier-eslint`
etc. to the latest compatible with what's in package.json:

  $ yarn upgrade --latest prettier
  $ yarn upgrade $(<package.json jq '.devDependencies | keys[]' -r |
                     grep ^prettier)

Then updated our formatting to match, with `tools/test --fix`:

  $ tools/test --full prettier --fix

Some changes are clear fixes; others seem more like just churn,
but *shrug*; easiest to go along.  In the parent commit, we added
ignores to block the few changes that were outright broken.