edit-message: Fix CANCEL_EDIT_MESSAGE test to actually work.
This test has never had its intended effect since it was first added,
in 5c31a466d.  I noticed it just now when making the lint fix in the
parent commit; now fix it.

Instead of the input this test is intended to pass, it passes an
unrelated, ill-typed, garbage input -- most likely due to a hasty
copy-paste from the test above it.

The app code, naturally, returns equally-garbage output.

Then the test cheerfully accepts the garbage.  In fact, it reproduces
the garbage output in its text, and checks for an exact match!
Naturally this garbage has little in common with what the app code
actually produces, and should produce, in the scenario the test is
supposedly testing.

The fact that the test checks the output against garbage does not
reflect well on the testing practices that produced it.  It's evident
that the expected output was written backward -- copied from the
actual output -- rather than based on any thinking about how the code
was actually intended to behave.

One likely root cause for this box-ticking testing practice and
garbled result is that these tests are painfully verbose and full of
boilerplate.  When it takes 20 lines of unrelated noise to express
this tiny test with its 2-3 lines of actual content, it's hard to
sustain attention, creatively write comprehensive tests, or catch
even blatant errors.

Another factor which certainly contributed is the fact that this test
code is untyped.  If the type-checker were active on this code, it
would have immediately flagged both the input and expected output as
ill-typed, and the bogus test would have been promptly fixed.

Both those factors are widespread in our tests, and were motivating
reasons for the recent `exampleData` module of test helpers.  Perhaps
I'll come back to this file in a bit and make it well-typed and less
verbose.  Anyway, for now, just make the test a proper test.