android build: Update Gradle to 5.4.1.
This is the version RN upstream uses as of v0.59.10.

Also include a comment saying how to upgrade.  I've had remarks in
commit messages 6dcb57357 and a1c9f1b77... but just now I first read
the former, earlier instruction; followed that; then saw it
accidentally moved us to "bin" and went looking to find the later one
to fix it.  Better to have it in the tree.

Release notes from Gradle upstream:
  https://docs.gradle.org/5.2.1/release-notes.html
  https://docs.gradle.org/5.3.1/release-notes.html
  https://docs.gradle.org/5.4.1/release-notes.html

They're fairly short, and nothing jumps out as affecting us.
Several changes that would be interesting if writing Gradle plugins,
or that would be helpful if switching to the Kotlin DSL.