babel: Move all config to babel.config.js .
We'd had both this file and a .babelrc .  The reason for that was
to accommodate Babel 6 and Babel 7 simultaneously, as described in
b32fc14b5.

But then in 529507336 we upgraded RN to v0.57, and put Babel 7-only
config syntax into our .babelrc .  It's not clear why we kept a
.babelrc around at all at that point, though from the commit message
it sounds like we did explicitly consider whether to do so.

Anyway, since facebook/react-native@5218932b1, the upstream RN
template puts this bit of config into babel.config.js , and has no
.babelrc .  Follow that.