flow: Upgrade to v0.81.
I was curious how far we could upgrade before finding errors
in RN itself -- this is the answer.  Further upgrades will
go along with the upcoming RN upgrade.

Found with a quick bisection.  Used the following command:

  $ v=0.81.0; \
  VER=$v perl -i -0pe '
      s/^\[version\]\n\K.*/^$ENV{"VER"}/m
    ' .flowconfig \
  && VER=$v perl -i -0pe '
      s/"flow-bin": \K.*/"^$ENV{"VER"}",/
    ' package.json \
  && yarn \
  && tools/test flow

but first tried 0.92.0 (the version used by RN v0.59), saw that fail,
then tried 0.85.0 and so on in a binary search.