webview: Upgrade react-native-webview to `^5.0.0`.
Fixes #3550.  This is also expected to fix #3518.

Release notes:
  https://github.com/react-native-community/react-native-webview/releases/tag/v3.0.0
  https://github.com/react-native-community/react-native-webview/releases/tag/v4.0.0
  https://github.com/react-native-community/react-native-webview/releases/tag/v5.0.0

In v5 in particular, there were breaking changes to communication from
inside the WebView to outside and vice versa.  In one direction, we can
follow the release notes: switch to `window.ReactNativeWebView.postMessage`.

In the other direction, upstream changed the API (seemingly by accident?)
so that, on iOS only, inbound events are dispatched at `window` rather
than `document`:
```
-￼   stringWithFormat:@"document.dispatchEvent(new MessageEvent('message', %@));",
+    stringWithFormat:@"window.dispatchEvent(new MessageEvent('message', %@));",
```
(from react-native-community/react-native-webview@f3bdab5a2 .)

So, add a hack to detect that and adjust accordingly.

[greg: expanded commit message; silenced lint error]