flow: Remove old fixme eliminated by upgrade.
As the `$FlowFixMe-56` indicates, this was introduced in the upgrade
of RN to v0.56 (and corresponding upgrade of Flow)... and the comment
doesn't sound like we had much idea at the time what the error
actually meant.

Now with the upcoming upgrade to RN v0.59 and of Flow to v0.92, the
error this applies to disappears.  Its disappearance seems just as
mysterious as its presence... it could be that it was just a bug and
there never was a type error, or it could be that the line the error
is reported on has moved around so that it's covered by a different
fixme comment, or something else.

Well, drop the fixme.

This also lets us re-enable failing on unused fixmes.