flow: Update libdefs.
This is partly a followup task from the RN v0.59 upgrade, #3399;
we've upgraded a lot of dependencies, so we should upgrade the
libdefs to match.

Done with:

  $ yarn update-libdefs

This turned up a couple of type errors, because there's now a real
libdef for WebView!  Marked those with fixmes; will fix properly next.