yarn: Silence warning from a peer-dep in react-intl.
Running `yarn` gives the following warning:

  warning " > react-intl@2.9.0" has unmet peer dependency "prop-types@^15.5.4".

(No such warning if `yarn` stops after step "Resolving packages" and
prints "Already up-to-date"; but if it has any actual work to do, it
always shows this warning in step "Linking dependencies".)

That's the only warning we now routinely get from `yarn`, so let's
get all the way to warning-free.

I believe the proper fix is that this shouldn't be a peer-dependency
in `react-intl` at all; reported that as formatjs/react-intl#1382 .
But as a workaround, adding a direct dependency silences the warning.