emoji/data: Make `nameToEmojiMap` beautiful.
I think this is rather more functional-looking than the old version
that we removed in the parent commit!  And this time genuinely
functional and not just abusing `reduce`.  ;-)

These would just say `Object.fromEntries(...)` -- using a function
quite recently added to the JS standard -- except that I'm not sure
our JS implementations have that, and in any case Flow doesn't believe
it exists.  So instead extract the common loop structure into a little
backport of `Object.fromEntries`.