flow: Explicitly type some literal `{}` to avoid losing type-checking.
In the common pattern of constructing an object-as-map by initializing
an empty object like `{}`, then adding some entries to it, Flow has an
odd behavior of deciding the type is `{||}`.

Seems like a bug in Flow.  The result is that if we put totally bogus
values into the object, Flow doesn't see any type error.  The fix,
or workaround, is to give the initial value an explicit type.

I noticed this when going through recently and converting all our
instances of that pattern from using `reduce` (a "false `reduce`")
to using a `forEach` loop.  The Flow issue applies in either form,
but this caused me to look at all the code that was in the old form.
Most already had this workaround, but some didn't, and a grep turned
up some already in the proper loop form that didn't either.

Fix them all, except one case where we can do even better by
converting to use the new `objectFromEntries`; we'll take care of that
in the next commit.