internal links: Pull out parsing of `pm-with` operands.
We'll do something parallel shortly for stream and topic/subject
operands, the better to contain the complexity we need to add to
stream operands in particular.

At the same time, we rewrite the logic somewhat to avoid using
NULL_USER.  The commit 2aa6c2264 earlier today already fixed it
so we never actually return bogus data taken from NULL_USER --
it's just a sentinel.  But `null` is a better sentinel, both
more explicit and with less risk of accidentally using it as
if it were real data.