nav: Use react-navigation-tabs for some features moved out to there.
This resolves the remaining three warnings introduced by our upgrade
to react-navigation v2:

    console.warn node_modules/react-navigation/src/react-navigation.js:63
      TabNavigator is deprecated. Please use the createBottomTabNavigator
      or createMaterialTopTabNavigator instead.
    console.warn node_modules/react-navigation/src/react-navigation.js:174
      TabBarTop is deprecated. Please use the react-navigation-tabs package
      instead: https://github.com/react-navigation/react-navigation-tabs
    console.warn node_modules/react-navigation/src/react-navigation.js:180
      TabBarBottom is deprecated. Please use the react-navigation-tabs package
      instead: https://github.com/react-navigation/react-navigation-tabs

For react-navigation-tabs, we pull in the latest version whose
peer-dependency on react-navigation points at v2 rather than v3:

  $ yarn add react-navigation-tabs@0.8.4

That got a peer-dep warning on react-native-screens.  So:

  $ yarn add react-native-screens@1.0.0

Turns out there isn't a post-alpha version :-/, but settled for
1.0.0-alpha.23 .

Then

  $ yarn update-libdefs

and then switched the code to use it!

This also eliminates some FlowFixMe comments... because the new
package has no meaningful libdef, so there's nothing to check.
Ah well.